% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footdist.R
\name{foot.dist}
\alias{foot.dist}
\title{Spearman's Footrule Distance Calculation}
\usage{
foot.dist(list1, list2)
}
\arguments{
\item{list1}{Ordered vector to compare}

\item{list2}{Ordered vector to compare}
}
\description{
Spearman's Footrule Distance Calculation
}
