% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCEL.R
\name{importCEL}
\alias{importCEL}
\title{Import CEL files and Export Matches}
\usage{
importCEL(phenoData, ...)
}
\arguments{
\item{phenoData}{AnnotatedDataFrame from read.AnnotatedDataFrame}

\item{...}{character vector of CEL file directory}
}
\value{
data.frame (tbl_df)
}
\description{
Import CEL files in the working directory and return the Quantile-Normalized
perfect matches!
}

