% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notNAs.R
\name{notNAs}
\alias{notNAs}
\title{Spearman's Footrule Distance Calculation}
\usage{
notNAs(vector)
}
\arguments{
\item{vector}{a vector from which you would like to remove NA values}
}
\description{
Spearman's Footrule Distance Calculation
}
